+++ homeassistant/components/matter/models.py
@@ -2,7 +2,6 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import TypedDict
 
 from chip.clusters import Objects as clusters
 from chip.clusters.Objects import ClusterAttributeDescriptor
@@ -15,20 +14,6 @@ from homeassistant.helpers.entity import EntityDescription
 SensorValueTypes = type[
     clusters.uint | int | clusters.Nullable | clusters.float32 | float
 ]
-
-
-class MatterDeviceInfo(TypedDict):
-    """Dictionary with Matter Device info.
-
-    Used to send to other Matter controllers,
-    such as Google Home to prevent duplicated devices.
-
-    Reference: https://developers.home.google.com/matter/device-deduplication
-    """
-
-    unique_id: str
-    vendor_id: str  # vendorId hex string
-    product_id: str  # productId hex string
 
 
 @dataclass
